Delete readme
**United Sensors of America**

We used Arduino and SnappyXO kit and a number of sensors (Ultrasonic, Temperature, Humidity, LDR) to create a robot able to solve the maze challenge, as well as explore the environment and interact with it and the members of family. We used Raspberry-Pi and its camera to run Face Detection algorithm that distinguishes between different family members, as well as detects unknown people and activates an alert. Also, we implemented mapping of sensors' data into Unity and its visualization in real time. For the Health Care bot, the app engine, Google Cloud storage, Softheon workshop UI, DialogFlow, Firebase were used - to create a personalized care-receiving experience.